requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
openai
sqlalchemy
pgvector
pypdf
psycopg[binary]
streamlit